refactor: remove `SiteGroup` To follow the same structure as bench and server list.
